services/alerts/requirements.txt
flask
kafka-python
python-dotenv
prometheus-client
